[torch.fx] Add test showing issue
Relying on fn.__code__.co_name, fn.__code__ checks limits the kinds of
functions that can be used inside a call to wrap, and also breaks when
used against decorated functions (fn.__name__ is updated to be the one
that's stored in the global dict, but fn.__code__.co_name is the one
that's defined in the wrapper).

Adding a unit test to explicitly demonstrate the issue.